Extending Your Todo
Added support for priority.
Added additional attribute in the listView like details, display
priority and a check box which can tell todo item is completed.
Added support for the SQLite. todo list is persisted all the times,
including its completed status.
Try adding a background color of the completed todos but there was some
bug, comment the code for now.
Reused the edit_item activity to add an item as well. (Need to clean.)